testbed-00/packages/epanet-fiware/requirements.txt
flake8==3.7.9
pytest==5.4.1
pytest-cov==2.7.1
numpy==1.20.3
pandas
matplotlib~=3.2.0
pyproj
requests
scikit-build
cmake
owa-epanet==2.2
click
pytest

